[Refactor] Home page 컨텐츠 pages로 옮기기
