+++ src/pages/Home/RankingContent.tsx
@@ -1,6 +1,6 @@
 import { useGetProductRankingList } from '@/hooks/query/product';
-import ProductInfiniteList from '../Product/ProductInfiniteList';
-import { ProductItem } from '../Product';
+import ProductInfiniteList from '../../components/Product/ProductInfiniteList';
+import { ProductItem } from '../../components/Product';
 
 export default function RankingContent() {
     const result = useGetProductRankingList('option');
